+++ CHANGELOG.md
@@ -9,6 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.0.0] - 2022-07-25
 
+### Features
+
+- Initial Commit
+- Create ```Azure Firewall```
+
 [Unreleased]: https://github.com/patrickhayo/azr-tf-module-firewall/compare/1.0.0...HEAD
 
 [1.0.0]: https://github.com/patrickhayo/azr-tf-module-firewall/compare/ce255d0522f468cac4f0494a68bf41e66c02eb87...1.0.0